requirements.txt
APScheduler==3.10.4
blinker==1.8.2
boto3==1.34.140
botocore==1.34.140
click==8.1.7
colorama==0.4.6
DateTime==5.5
Flask==3.0.3
Flask-Cors==4.0.1
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
MarkupSafe==2.1.5
numpy==2.0.0
packaging==24.1
pandas==2.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
s3transfer==0.10.2
scikit-learn==1.5.1
scipy==1.14.0
setuptools==70.2.0
six==1.16.0
threadpoolctl==3.5.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
waitress==3.0.0
Werkzeug==3.0.3
zope.interface==6.4.post2
